pagesリファクタリング
