+++ src/components/layouts/DashBoard.tsx
@@ -1,174 +1,16 @@
-import * as React from "react";
-import { styled, useTheme } from "@mui/material/styles";
+import { NextPage } from "next";
 import Box from "@mui/material/Box";
-import Drawer from "@mui/material/Drawer";
-import CssBaseline from "@mui/material/CssBaseline";
-import MuiAppBar, { AppBarProps as MuiAppBarProps } from "@mui/material/AppBar";
-import Toolbar from "@mui/material/Toolbar";
-
-import Typography from "@mui/material/Typography";
-import IconButton from "@mui/material/IconButton";
-import MenuIcon from "@mui/icons-material/Menu";
-import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import ChevronRightIcon from "@mui/icons-material/ChevronRight";
-import ListItem from "@mui/material/ListItem";
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
-import LibraryMusicIcon from "@mui/icons-material/LibraryMusic";
-import LyricsIcon from "@mui/icons-material/Lyrics";
-import AccountBoxIcon from "@mui/icons-material/AccountBox";
-import HomeIcon from "@mui/icons-material/Home";
-import Link from "next/link";
-const drawerWidth = 240;
-
-interface AppBarProps extends MuiAppBarProps {
-  open?: boolean;
-}
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== "open",
-})<AppBarProps>(({ theme, open }) => ({
-  transition: theme.transitions.create(["margin", "width"], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    width: `calc(100% - ${drawerWidth}px)`,
-    marginLeft: 0,
-    transition: theme.transitions.create(["margin", "width"], {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
-
-const DrawerHeader = styled("div")(({ theme }) => ({
-  display: "flex",
-  alignItems: "center",
-  padding: theme.spacing(0, 1),
-  // necessary for content to be below app bar
-  ...theme.mixins.toolbar,
-  justifyContent: "flex-end",
-}));
+import Header from "../elements/Header";
 
 type LayoutProps = {
   children: React.ReactNode;
 };
 
-export default function DashBoard({ children }: LayoutProps) {
-  const theme = useTheme();
-  const [open, setOpen] = React.useState(false);
-
-  const handleDrawerOpen = () => {
-    setOpen(true);
-  };
-
-  const handleDrawerClose = () => {
-    setOpen(false);
-  };
-
+export const DashBoard: NextPage<LayoutProps> = (props) => {
   return (
     <Box sx={{ display: "flex" }}>
-      <CssBaseline />
-      <AppBar position="fixed" open={open}>
-        <Toolbar>
-          <IconButton
-            color="inherit"
-            aria-label="open drawer"
-            onClick={handleDrawerOpen}
-            edge="start"
-            sx={{ mr: 2, ...(open && { display: "none" }) }}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Link href="/" style={{ color: "white", textDecoration: "none" }}>
-            <Typography variant="h6" noWrap component="div">
-              B e a t O r
-            </Typography>
-          </Link>
-        </Toolbar>
-      </AppBar>
-      <Drawer
-        sx={{
-          width: drawerWidth,
-          flexShrink: 0,
-          "& .MuiDrawer-paper": {
-            width: drawerWidth,
-            boxSizing: "border-box",
-          },
-        }}
-        variant="temporary"
-        anchor="left"
-        open={open}
-      >
-        <DrawerHeader>
-          <IconButton onClick={handleDrawerClose}>
-            {theme.direction === "ltr" ? (
-              <ChevronLeftIcon />
-            ) : (
-              <ChevronRightIcon />
-            )}
-          </IconButton>
-        </DrawerHeader>
-
-        <ListItem disablePadding>
-          <Link href="/" style={{ color: "black", textDecoration: "none" }}>
-            <ListItemButton>
-              <ListItemIcon>
-                <HomeIcon />
-              </ListItemIcon>
-              <ListItemText primary="ホーム画面" />{" "}
-            </ListItemButton>
-          </Link>
-        </ListItem>
-
-        {/* 楽曲一覧のリストアイテム */}
-        <ListItem disablePadding>
-          <Link
-            href="/song-list"
-            style={{ color: "black", textDecoration: "none" }}
-          >
-            <ListItemButton>
-              <ListItemIcon>
-                <LibraryMusicIcon /> {/* 楽曲一覧アイコン */}
-              </ListItemIcon>
-              <ListItemText primary="楽曲一覧" /> {/* テキスト: "楽曲一覧" */}
-            </ListItemButton>
-          </Link>
-        </ListItem>
-
-        <ListItem disablePadding>
-          <Link
-            href="/create-song"
-            style={{ color: "black", textDecoration: "none" }}
-          >
-            <ListItemButton>
-              <ListItemIcon>
-                <LyricsIcon /> {/* 楽曲新規制作:アイコン */}
-              </ListItemIcon>
-              <ListItemText primary="楽曲新規制作" />{" "}
-              {/* テキスト: "楽曲新規制作" */}
-            </ListItemButton>
-          </Link>
-        </ListItem>
-
-        <ListItem disablePadding>
-          <Link
-            href="/profile"
-            style={{ color: "black", textDecoration: "none" }}
-          >
-            <ListItemButton>
-              <ListItemIcon>
-                <AccountBoxIcon /> {/* プロフィールアイコン */}
-              </ListItemIcon>
-              <ListItemText primary="プロフィール" />{" "}
-              {/* テキスト: "プロフィール" */}
-            </ListItemButton>
-          </Link>
-        </ListItem>
-      </Drawer>
-      <Box sx={{ marginTop: "64px", width: "100%" }}>{children}</Box>
+      <Header />
+      <Box sx={{ marginTop: "64px", width: "100%" }}>{props.children}</Box>
     </Box>
   );
-}
+};
