+++ src/pages/index.tsx
@@ -1,11 +1,9 @@
-import { Inter } from "next/font/google";
-import DashBoard from "@/components/layouts/DashBoard";
+import { NextPage } from "next";
+import { DashBoard } from "@/components/layouts/DashBoard";
 import SoundControlBar from "@/features/audioPlayer/components/SoundControlBar";
 import LookingForSounds from "@/features/lookingForSounds/components/LookingForSounds";
 
-const inter = Inter({ subsets: ["latin"] });
-
-export default function Home() {
+const Home: NextPage = () => {
   return (
     <main>
       <DashBoard>
@@ -14,4 +12,6 @@ export default function Home() {
       </DashBoard>
     </main>
   );
-}
+};
+
+export default Home;
