+++ src/pages/song-list/index.tsx
@@ -1,17 +1,14 @@
-import { Inter } from "next/font/google";
-import DashBoard from "@/components/layouts/DashBoard";
+import { NextPage } from "next";
+import { DashBoard } from "@/components/layouts/DashBoard";
 import SoundControlBar from "@/features/audioPlayer/components/SoundControlBar";
 
-
-const inter = Inter({ subsets: ["latin"] });
-
-export default function Home() {
+export const SongList: NextPage = () => {
   return (
     <main>
       <DashBoard>
         楽曲一覧画面
-       <SoundControlBar />
+        <SoundControlBar />
       </DashBoard>
     </main>
   );
-}
+};
